Flatten conductivity sensor UI into single script
